tests/src/requirements/bert_requirements.txt
azure-ai-ml
azureml-core
azureml-mlflow
transformers
transformers[torch]
torchvision
mlflow
SentencePiece
evaluate